[Feat] 리뷰 조회 쿼리 최종 수정 후 테스트 코드 모두 수정

+++ src/test/java/com/inq/wishhair/wesharewishhair/review/domain/ReviewTest.java
@@ -48,7 +48,7 @@ public class ReviewTest {
         Review review = ReviewFixture.A.toEntity(user, hairStyle);
 
         //when
-        int likes = review.getLikes();
+        long likes = review.getLikes();
 
         //then
         assertThat(likes).isZero();
